// src/app/interfaces/question.interface.ts
export interface Question {
  title: string;
  text: string;
  tags: Array<string>;
}

// Generated by https://quicktype.io

export interface QuestionResponse {
  _id: string;
  author: Author;
  title: string;
  text: string;
  tags: string[];
  views: number;
  votes: any[];
  comments: any[];
  answers: any[];
  createdAt: string;
  updatedAt: string;
}

export interface Author {
  _id: string;
  username: string;
  email: string;
  createdAt: string;
  updatedAt: string;
}